requirements.txt
paramiko>=2.12.0
requests>=2.31.0
colorama>=0.4.6
PySocks>=1.7.1
urllib3>=2.1.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
tqdm>=4.66.1
cryptography>=41.0.7
aiohttp>=3.9.1
python-whois>=0.8.0
dnspython>=2.4.2
shodan>=1.30.1
censys>=2.2.8
pdfkit>=1.0.0
jinja2>=3.1.2
python-dateutil>=2.8.2
pyOpenSSL>=23.3.0
stem>=1.8.2
validators>=0.22.0
python-magic>=0.4.27
pycryptodome>=3.19.0
PyYAML>=6.0.1
python-nmap>=0.7.1
selenium
webdriver-manager